:bug: Fixes removal of dylibs post codesign = unable to build for AppStore from VS
As Xamarin for Visual Studio is missing the 'Sign and Distribute' function in Xamarin Studio, the ipa is not resigned after unneeded dylibs are removed. 

The solution now contains both iPhone and iPhoneSImulator configurations which allows only the dylibs required to be bundled into the dll. Also added a targets file that automatically injects the reference on build based on configuration.